+++ lib/mpv_audio_kit.dart
@@ -31,8 +31,10 @@ export 'src/models/audio_filter.dart' show AudioFilter;
 export 'src/models/audio_params.dart' show AudioParams;
 export 'src/models/mpv_log_entry.dart' show MpvLogEntry;
 export 'src/models/mpv_hook_event.dart' show MpvHookEvent;
+export 'src/models/mpv_player_error.dart'
+    show MpvPlayerError, MpvEndFileError, MpvLogError, MpvEndFileReason, MpvFileEndedEvent;
 export 'src/models/player_configuration.dart' show PlayerConfiguration;
 export 'src/models/player_state.dart' show PlayerState;
 export 'src/models/player_stream.dart' show PlayerStream;
-export 'src/mpv_bindings.dart' show MpvLibraryException;
+export 'src/mpv_bindings.dart' show MpvLibraryException, MpvError;
 export 'src/mpv_audio_kit.dart' show MpvAudioKit;